Use S3 for plotting, give more structure to output
We now class the output of gprd(), and to make plots, just use a plot()
method. Additionally, we return a lot more from the C++ calculations,
so that predicting future cases is quicker after having run the model
once. Using an S3 class will also be useful (we can add a summary()
method, etc....)