Add summary method, new sanity check, fix CI names
First, in the documentation, I said I would name the elements of the CI
vector, which seems like a good idea, but then for some reason I did
not. This commit fixes that.

Second, I add a sanity check to ensure the CI width is in [0, 1].

Finally, and most substantially, I added a summary method. It is pretty
simple, just printing the treatment effect estimate and CI in 32
character width with a header.