Rename local estimator to piecewise, fix R CMD check issues
Joe came up with a much better name for what we had been calling the
'local' estimator: the 'piecewise' estimator. So, I changed that name
everywhere. For now, we fail gracefully if 'local' is specified,
changing it to local with a warning. I'll remove that bit soon enough.

I also found out through R CMD check that there were some more stats
and graphics functions that I had not yet setup NAMESPACE imports for,
so I fixed those. Additionally, the check found a spot where I had
misspelled a variable (how did I miss that??), so that was fixed.
Finally, I didn't set up the formals for the summary() method quite
right, so I fixed that.